refactor: separate test utils,constants for readability
